Fix project tests.
